# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: arc — Personal Brand Strategy API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /dossier:
    get:
      operationId: getDossier
      tags: [dossier]
      summary: Get the latest briefing dossier
      responses:
        "200":
          description: Briefing dossier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BriefingDossier"
        "404":
          description: Not found
  /dossier/generate:
    post:
      operationId: generateDossier
      tags: [dossier]
      summary: Research the client and competitors into a briefing dossier
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DossierGenerateInput"
      responses:
        "200":
          description: Generated briefing dossier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BriefingDossier"
        "404":
          description: No client profile yet
        "502":
          description: Generation failed
...
  /industry-overview:
    get:
      operationId: getIndustryOverview
      tags: [industryOverview]
      summary: Get the latest industry overview
      responses:
        "200":
          description: Industry overview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IndustryOverview"
        "404":
          description: Not found
  /industry-overview/generate:
    post:
      operationId: generateIndustryOverview
      tags: [industryOverview]
      summary: Research the client's industry landscape into an industry overview
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IndustryOverviewGenerateInput"
      responses:
        "200":
          description: Generated industry overview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IndustryOverview"
        "403":
          description: Foundation incomplete
        "404":
          description: No client profile yet
        "502":
          description: Generation failed
  /content-strategy:
    get:
      operationId: getContentStrategy
      tags: [contentStrategy]
      summary: Get the content strategy
      responses:
        "200":
          description: Content strategy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentStrategy"
        "404":
          description: Not found
    put:
      operationId: updateContentStrategy
      tags: [contentStrategy]
      summary: Update the editable text fields of the content strategy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContentStrategyUpdateInput"
      responses:
        "200":
          description: Updated content strategy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentStrategy"
        "404":
          description: Not found
  /content-strategy/generate:
    post:
      operationId: generateContentStrategy
      tags: [contentStrategy]
      summary: Generate a content strategy from the blueprint and platform strategy
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContentStrategyGenerateInput"
      responses:
        "200":
          description: Generated content strategy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentStrategy"
        "403":
          description: Blueprint or platform strategy incomplete
        "404":
          description: No client profile yet
        "502":
          description: Generation failed
...
  /ideas:
    get:
      operationId: listIdeas
      tags: [ideas]
      summary: List content ideas
      responses:
        "200":
          description: List of ideas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Idea"
    post:
      operationId: createIdea
      tags: [ideas]
      summary: Create a content idea
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IdeaInput"
      responses:
        "201":
          description: Created idea
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Idea"
  /ideas/{id}:
    delete:
      operationId: deleteIdea
      tags: [ideas]
      summary: Delete a content idea
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
    patch:
      operationId: updateIdea
      tags: [ideas]
      summary: Update a content idea
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IdeaUpdate"
      responses:
        "200":
          description: Updated idea
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Idea"
...
components:
  schemas:
...
    Competitor:
      type: object
      required: [name, description, positioning, differentiation]
      properties:
        name:
          type: string
        description:
          type: string
        positioning:
          type: string
        differentiation:
          type: string
    DossierSource:
      type: object
      required: [title, url]
      properties:
        title:
          type: string
        url:
          type: string
    BriefingDossier:
      type: object
      required: [id, clientId, footprintSummary, competitors, sources, generatedAt, createdAt]
      properties:
        id:
          type: integer
        clientId:
          type: integer
        footprintSummary:
          type: string
        competitors:
          type: array
          items:
            $ref: "#/components/schemas/Competitor"
        sources:
          type: array
          items:
            $ref: "#/components/schemas/DossierSource"
        generatedAt:
          type: string
        createdAt:
          type: string
    DossierGenerateInput:
      type: object
      properties:
        feedback:
          type: string
          description: Optional user revision notes to steer a regeneration.
...
    IndustryPlayer:
      type: object
      required: [name, description, positioning]
      properties:
        name:
          type: string
        description:
          type: string
        positioning:
          type: string
    PlaybookMove:
      type: object
      required: [title, detail]
      properties:
        title:
          type: string
        detail:
          type: string
    IndustrySource:
      type: object
      required: [title, url]
      properties:
        title:
          type: string
        url:
          type: string
    IndustryOverview:
      type: object
      required:
        [
          id,
          clientId,
          industry,
          geographyFocus,
          landscapeContext,
          competitors,
          thoughtLeaders,
          playbook,
          sources,
          generatedAt,
          createdAt,
        ]
      properties:
        id:
          type: integer
        clientId:
          type: integer
        industry:
          type: string
        geographyFocus:
          type: string
        landscapeContext:
          type: string
        competitors:
          type: array
          items:
            $ref: "#/components/schemas/IndustryPlayer"
        thoughtLeaders:
          type: array
          items:
            $ref: "#/components/schemas/IndustryPlayer"
        playbook:
          type: array
          items:
            $ref: "#/components/schemas/PlaybookMove"
        sources:
          type: array
          items:
            $ref: "#/components/schemas/IndustrySource"
        generatedAt:
          type: string
        createdAt:
          type: string
    IndustryOverviewGenerateInput:
      type: object
      properties:
        feedback:
          type: string
          description: Optional user revision notes to steer a regeneration.
    PlatformCadence:
      type: object
      required: [platform, frequency, formats, focus]
      properties:
        platform:
          type: string
        frequency:
          type: string
        formats:
          type: array
          items:
            type: string
        focus:
          type: string
    ContentMixItem:
      type: object
      required: [type, description, whyForClient, exampleTopics, weight]
      properties:
        type:
          type: string
        description:
          type: string
        whyForClient:
          type: string
        exampleTopics:
          type: array
          items:
            type: string
        weight:
          type: string
    SignatureSeries:
      type: object
      required: [name, cadence, description]
      properties:
        name:
          type: string
        cadence:
          type: string
        description:
          type: string
    PostFormat:
      type: object
      required: [name, description]
      properties:
        name:
          type: string
        description:
          type: string
    ContentStrategy:
      type: object
      required:
        [
          id,
          clientId,
          summary,
          platformPlan,
          contentMix,
          signatureSeries,
          postFormats,
          repurposing,
          closing,
          createdAt,
          updatedAt,
        ]
      properties:
        id:
          type: integer
        clientId:
          type: integer
        summary:
          type: string
        platformPlan:
          type: array
          items:
            $ref: "#/components/schemas/PlatformCadence"
        contentMix:
          type: array
          items:
            $ref: "#/components/schemas/ContentMixItem"
        signatureSeries:
          type: array
          items:
            $ref: "#/components/schemas/SignatureSeries"
        postFormats:
          type: array
          items:
            $ref: "#/components/schemas/PostFormat"
        repurposing:
          type: string
        closing:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
...
    ContentStrategyGenerateInput:
      type: object
      properties:
        feedback:
          type: string
          description: Optional user revision notes to steer a regeneration.
...
    Idea:
      type: object
      required: [id, title, notes, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        notes:
          type: string
        platform:
          type: ["string", "null"]
        createdAt:
          type: string
    IdeaInput:
      type: object
      required: [title]
      properties:
        title:
          type: string
        notes:
          type: string
        platform:
          type: string
    IdeaUpdate:
      type: object
      properties:
        title:
          type: string
        notes:
          type: string
        platform:
          type: string
...
    ContentStrategyUpdateInput:
      type: object
      required: [summary, repurposing, closing]
      properties:
        summary:
          type: string
        repurposing:
          type: string
        closing:
          type: string
...